assignment/requirements.txt
firebase_admin==6.5.0
Flask==3.0.3
google_auth_oauthlib==1.2.1
PyJWT==2.9.0
Requests==2.32.3
gunicorn==20.1.0